feat: Added mutation upload and bullet proof other commands

+++ src/gpaslocal/importer.py
@@ -41,17 +41,18 @@ def import_data(excel_wb: str, dryrun: bool = False) -> bool:
         except Exception as e:
             logger.error(f"Failed to upload data: {e}")
 
-        if logger.error_occurred:  # type: ignore
-            session.rollback()
-            logger.error("Upload failed, please see log messages for details")
-            return False
+        finally:
+            if logger.error_occurred:  # type: ignore
+                session.rollback()
+                logger.error("Upload failed, please see log messages for details")
+                return False
 
-        if dryrun:
-            logger.info("Dry run mode, no data was uploaded")
-            session.rollback()
-        else:
-            logger.info("Data uploaded successfully")
-            session.commit()
+            if dryrun:
+                logger.info("Dry run mode, no data was uploaded")
+                session.rollback()
+            else:
+                logger.info("Data uploaded successfully")
+                session.commit()
 
     return True
 
@@ -64,24 +65,25 @@ def runs(session: Session, excel_wb: str, dryrun: bool) -> None:
         try:
             run_import = RunImport(**row)
 
-            if run_record := (
+            run_record = (
                 session.query(models.Run)
                 .filter(models.Run.code == run_import.code)
                 .first()
-            ):
-                # update the run record
+            )
+            if run_record:
                 run_record.update_from_importmodel(run_import)
                 logger.info(
                     f"Runs Sheet Row {index+2}: Run {run_import.code} already exists{'' if dryrun else ', updating'}"
                 )
             else:
                 # add the run record
-                run_record = models.Run()
-                run_record.update_from_importmodel(run_import)
+                run_record = models.Run(code=run_import.code)
                 session.add(run_record)
                 logger.info(
                     f"Runs Sheet Row {index+2}: Run {run_import.code} does not exist{'' if dryrun else ', adding'}"
                 )
+            run_record.update_from_importmodel(run_import)
+
         except ValidationError as err:
             for error in err.errors():
                 logger.error(
@@ -102,27 +104,31 @@ def specimens(session: Session, excel_wb: str, dryrun: bool) -> None:
             # get the specimen owner
             owner_record = owner(session, index, row, dryrun)
 
-            if specimen_record := (
+            specimen_record = (
                 session.query(models.Specimen)
                 .filter(
                     models.Specimen.accession == specimen_import.accession,
                     models.Specimen.collection_date == specimen_import.collection_date,
                 )
                 .first()
-            ):
-                specimen_record.update_from_importmodel(specimen_import)
-                specimen_record.owner = owner_record
+            )
+
+            if specimen_record:
                 logger.info(
                     f"Specimens Sheet Row {index+2}: Specimen {specimen_import.accession}, {specimen_import.collection_date} already exists{'' if dryrun else ', updating'}"
                 )
             else:
-                specimen_record = models.Specimen()
-                specimen_record.update_from_importmodel(specimen_import)
-                specimen_record.owner = owner_record
+                specimen_record = models.Specimen(
+                    accession=specimen_import.accession,
+                    collection_date=specimen_import.collection_date,
+                )
                 session.add(specimen_record)
                 logger.info(
                     f"Specimens Sheet Row {index+2}: Specimen {specimen_import.accession}, {specimen_import.collection_date} does not exist{'' if dryrun else ', adding'}"
                 )
+            specimen_record.update_from_importmodel(specimen_import)
+            specimen_record.owner = owner_record
+            session.flush()
 
             specimen_detail(session, specimen_record, specimen_import)
 
@@ -165,31 +171,38 @@ def specimen_detail(
     for specimen_detail_type in specimen_detail_types:
         # get the value from the sample_import
         value = specimen_import[specimen_detail_type.code]
-        # check if the sample detail exists
+
+        # check if the specimen has been added to the database yet
+        # if so check if the specimen detail exists
         specimen_detail_record = (
-            session.query(models.SpecimenDetail)
-            .filter(
-                models.SpecimenDetail.specimen == specimen_record,
-                models.SpecimenDetail.specimen_detail_type_code
-                == specimen_detail_type.code,
+            (
+                session.query(models.SpecimenDetail)
+                .filter(
+                    models.SpecimenDetail.specimen == specimen_record,
+                    models.SpecimenDetail.specimen_detail_type_code
+                    == specimen_detail_type.code,
+                )
+                .first()
             )
-            .first()
+            if specimen_record.id
+            else None
         )
 
         # if the value is None, and the sample detail exists, delete it
+        # if the value is None skip
         if value is None:
             if specimen_detail_record:
                 session.delete(specimen_detail_record)
             continue
 
-        if specimen_detail_record:
-            specimen_detail_record["value_" + specimen_detail_type.value_type] = value
-        else:
-            specimen_detail_record = models.SpecimenDetail()
-            specimen_detail_record.specimen = specimen_record
-            specimen_detail_record.specimen_detail_type_code = specimen_detail_type.code
-            specimen_detail_record["value_" + specimen_detail_type.value_type] = value
+        if not specimen_detail_record:
+            specimen_detail_record = models.SpecimenDetail(
+                specimen=specimen_record,
+                specimen_detail_type_code=specimen_detail_type.code,
+            )
             session.add(specimen_detail_record)
+
+        specimen_detail_record["value_" + specimen_detail_type.value_type] = value
 
 
 def samples(session: Session, excel_wb: str, dryrun: bool) -> None:
@@ -228,6 +241,9 @@ def samples(session: Session, excel_wb: str, dryrun: bool) -> None:
                     f"Samples Sheet Row {index+2}: Sample {sample_import.guid} does not exist{'' if dryrun else ', adding'}"
                 )
                 session.add(sample_record)
+
+            # make sure we have the sample record id
+            session.flush()
 
             # add the sample detail records
             sample_detail(session, sample_record, sample_import)
@@ -275,30 +291,36 @@ def sample_detail(
     for sample_detail_type in sample_detail_types:
         # get the value from the sample_import
         value = sample_import[sample_detail_type.code]
+
         # check if the sample detail exists
         sample_detail_record = (
-            session.query(models.SampleDetail)
-            .filter(
-                models.SampleDetail.sample == sample_record,
-                models.SampleDetail.sample_detail_type_code == sample_detail_type.code,
+            (
+                session.query(models.SampleDetail)
+                .filter(
+                    models.SampleDetail.sample == sample_record,
+                    models.SampleDetail.sample_detail_type_code
+                    == sample_detail_type.code,
+                )
+                .first()
             )
-            .first()
+            if sample_record.id
+            else None
         )
 
         # if the value is None, and the sample detail exists, delete it
+        # if the value is None skip
         if value is None:
             if sample_detail_record:
                 session.delete(sample_detail_record)
             continue
 
-        if sample_detail_record:
-            sample_detail_record["value_" + sample_detail_type.value_type] = value
-        else:
-            sample_detail_record = models.SampleDetail()
-            sample_detail_record.sample = sample_record
-            sample_detail_record.sample_detail_type_code = sample_detail_type.code
-            sample_detail_record["value_" + sample_detail_type.value_type] = value
+        if not sample_detail_record:
+            sample_detail_record = models.SampleDetail(
+                sample=sample_record, sample_detail_type_code=sample_detail_type.code
+            )
             session.add(sample_detail_record)
+
+        sample_detail_record["value_" + sample_detail_type.value_type] = value
 
 
 def spikes(
@@ -337,28 +359,36 @@ def spikes(
             )
             continue
 
-        if spike_record := (
-            session.query(models.Spike)
-            .filter(
-                models.Spike.sample == sample_record,
-                models.Spike.name == spike_name,
+        spike_record = (
+            (
+                session.query(models.Spike)
+                .filter(
+                    models.Spike.sample == sample_record,
+                    models.Spike.name == spike_name,
+                )
+                .first()
             )
-            .first()
-        ):
-            spike_record.quantity = spike_quantity
-        else:
-            spike_record = models.Spike()
-            spike_record.sample = sample_record
-            spike_record.name = spike_name
-            spike_record.quantity = spike_quantity
+            if sample_record.id
+            else None
+        )
+
+        if not spike_record:
+            spike_record = models.Spike(
+                sample=sample_record,
+                name=spike_name,
+            )
             session.add(spike_record)
 
+        spike_record.quantity = spike_quantity
+
     # remove any spikes that are not in the spike table for this sample
-    clean_spike_names = [x for x in spike_names.values() if not pd.isnull(x)]
-    session.query(models.Spike).filter(
-        not_(models.Spike.name.in_(clean_spike_names)),
-        models.Spike.sample == sample_record,
-    ).delete()
+    # only if the sample has been added to the database
+    if sample_record.id:
+        clean_spike_names = [x for x in spike_names.values() if not pd.isnull(x)]
+        session.query(models.Spike).filter(
+            not_(models.Spike.name.in_(clean_spike_names)),
+            models.Spike.sample == sample_record,
+        ).delete()
 
 
 def storage(session: Session, excel_wb: str, dryrun: bool) -> None:
@@ -373,26 +403,33 @@ def storage(session: Session, excel_wb: str, dryrun: bool) -> None:
                 session, storage_import.accession, storage_import.collection_date
             )
 
-            if storage_record := (
-                session.query(models.Storage)
-                .filter(
-                    models.Storage.storage_qr_code == storage_import.storage_qr_code
+            storage_record = (
+                (
+                    session.query(models.Storage)
+                    .filter(
+                        models.Storage.storage_qr_code == storage_import.storage_qr_code
+                    )
+                    .first()
                 )
-                .first()
-            ):
-                storage_record.update_from_importmodel(storage_import)
-                storage_record.specimen = specimen_record
-                logger.info(
-                    f"Storage Sheet Row {index+2}: Storage {storage_import.storage_qr_code} already exists{'' if dryrun else ', updating'}"
+                if specimen_record.id
+                else None
+            )
+
+            if not storage_record:
+                storage_record = models.Storage(
+                    storage_qr_code=storage_import.storage_qr_code
                 )
-            else:
-                storage_record = models.Storage()
-                storage_record.update_from_importmodel(storage_import)
-                storage_record.specimen = specimen_record
+                session.add(storage_record)
                 logger.info(
                     f"Storage Sheet Row {index+2}: Storage {storage_import.storage_qr_code} does not exist{'' if dryrun else ', adding'}"
                 )
-                session.add(storage_record)
+            else:
+                logger.info(
+                    f"Storage Sheet Row {index+2}: Storage {storage_import.storage_qr_code} already exists{'' if dryrun else ', updating'}"
+                )
+
+            storage_record.update_from_importmodel(storage_import)
+            storage_record.specimen = specimen_record
 
         except ValidationError as err:
             for error in err.errors():
